+++ code_pylint.py
@@ -18,7 +18,7 @@ MIN_NOTE = 9.17
 UNWATCHED_ERRORS = ['fixme', 'trailing-whitespace', 'import-error']
 
 MAX_ERROR_BY_TYPE = {
-                     'protected-access': 40,
+                     'protected-access': 41,
                      'invalid-name': 4,
                      'consider-using-f-string': 1,
                      'no-else-return': 0,
@@ -43,7 +43,7 @@ MAX_ERROR_BY_TYPE = {
                      'raise-missing-from': 0,
                      'consider-merging-isinstance': 1,
                      'abstract-method': 6,
-                     'import-outside-toplevel': 11,  # TODO : will reduced in a future work (when tests are ready)
+                     'import-outside-toplevel': 12,  # TODO : will reduced in a future work (when tests are ready)
                      'too-many-instance-attributes': 7,
                      'consider-iterating-dictionary': 4,
                      'attribute-defined-outside-init': 3,